Add option to specify preferred binaries instead of system versions
This feature allows a different root directory to be specified, ideally
containing statically-linked versions of binaries to be used. If a
particular needed binary is not found in the specified alternate root
path, the system-installed version is used.

This also copies binaries to the initramfs verbosely to show which
specific binaries are being used.